% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sen_smooth.R
\name{sen}
\alias{sen}
\title{Thiel-Sen Smooth}
\usage{
sen(..., weights = NULL)
}
\arguments{
\item{...}{passes formula and dataframe to function}

\item{weights}{optional vector of weights}
}
\description{
Use mblm to calculate the Thiel-Sen smooth, which is the median of all slopes between all pairs of points with different x-values. See the documentation for mblm::mblm(repeated = FALSE) for full details. Use this as the method within a function, e.g. ggplot2::geom_smooth(method = "sen").
}
