% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sen_smooth.R
\name{siegel}
\alias{siegel}
\title{Siegel Smooth}
\usage{
siegel(..., weights = NULL)
}
\arguments{
\item{...}{passes formula and dataframe to function}

\item{weights}{optional vector of weights}
}
\description{
Use mblm to calculate the Siegel smooth, which is the nested median of all slopes. Generally recommended for 14 or more data points. See the documentation for mblm::mblm(repeated = TRUE) for full details. Use this as the method within a function, e.g. ggplot2::geom_smooth(method = "sen").
}
